+++ example/lib/main_wav_stream.dart
@@ -102,8 +102,10 @@ class _MyHomePageState extends State<MyHomePage> {
 
   Future<void> addSound(String file) async {
     var l = 0.0;
-    final s = await SoLoudTools.loadFromFile(file, mode: LoadMode.disk);
-    sounds.add(s!);
+    final ret = await SoLoud.instance.loadFile(file, mode: LoadMode.disk);
+    assert(ret.error == PlayerErrors.noError, 'Error: ${ret.error}');
+    final s = ret.sound!;
+    sounds.add(s);
     l = SoLoud.instance.getLength(s).length;
     if (l < minLength) minLength = l;
   }