+++ lib/src/bindings_player_ffi.dart
@@ -122,7 +122,7 @@ class FlutterSoLoudFfi {
   /// If `LoadMode.disk` is used, the audio data is loaded
   /// from the given file when needed (more CPU, less memory allocated).
   /// See the [seek] note problem when using [LoadMode] = `LoadMode.disk`.
-  /// [soundHash] return hash of the sound.
+  /// `soundHash` return hash of the sound.
   /// Returns [PlayerErrors.noError] if success.
   ({PlayerErrors error, SoundHash soundHash}) loadFile(
     String completeFileName,
@@ -158,7 +158,7 @@ class FlutterSoLoudFfi {
   /// [superWave]
   /// [scale]
   /// [detune]
-  /// [soundHash] return hash of the sound
+  /// `soundHash` return hash of the sound
   /// Returns [PlayerErrors.noError] if success
   ({PlayerErrors error, SoundHash soundHash}) loadWaveform(
     WaveForm waveform,
@@ -369,37 +369,50 @@ class FlutterSoLoudFfi {
   late final _getRelativePlaySpeed =
       _getRelativePlaySpeedPtr.asFunction<double Function(int)>();
 
-  /// Play already loaded sound identified by [soundHash].
+  /// Play already loaded sound identified by [soundHash]
   ///
   /// [soundHash] the unique sound hash of a sound
-  /// [volume] 1.0f full volume
-  /// [pan] 0.0f centered
-  /// [paused] 0 not pause
-  ///
-  /// Return the handle of the sound, [SoundHandle.error] if error
-  SoundHandle play(
+  /// [volume] 1.0 full volume
+  /// [pan] 0.0 centered
+  /// [paused] false not paused
+  /// [looping] whether to start the sound in looping state.
+  /// [loopingStartAt] If looping is enabled, the loop point is, by default,
+  /// the start of the stream. The loop start point can be set with this
+  /// parameter, and current loop point can be queried with [getLoopingPoint]
+  /// and changed by [setLoopingPoint].
+  /// Return the error if any and a new [newHandle] of this sound
+  ({PlayerErrors error, SoundHandle newHandle}) play(
     SoundHash soundHash, {
     double volume = 1,
     double pan = 0,
     bool paused = false,
+    bool looping = false,
+    double loopingStartAt = 0,
   }) {
-    final handleId = _play(soundHash.hash, volume, pan, paused ? 1 : 0);
-    if (handleId == 0) {
-      return SoundHandle.error();
-    }
-    return SoundHandle(handleId);
+    // ignore: omit_local_variable_types
+    final ffi.Pointer<ffi.UnsignedInt> handle = calloc();
+    final e = _play(
+      soundHash.hash,
+      volume,
+      pan,
+      paused ? 1 : 0,
+      looping ? 1 : 0,
+      loopingStartAt,
+      handle,
+    );
+    final ret =
+        (error: PlayerErrors.values[e], newHandle: SoundHandle(handle.value));
+    calloc.free(handle);
+    return ret;
   }
 
   late final _playPtr = _lookup<
       ffi.NativeFunction<
-          ffi.UnsignedInt Function(
-            ffi.UnsignedInt,
-            ffi.Float,
-            ffi.Float,
-            ffi.Int,
-          )>>('play');
-  late final _play =
-      _playPtr.asFunction<int Function(int, double, double, int)>();
+          ffi.Int32 Function(ffi.UnsignedInt, ffi.Float, ffi.Float, ffi.Int,
+              ffi.Int, ffi.Double, ffi.Pointer<ffi.UnsignedInt>)>>('play');
+  late final _play = _playPtr.asFunction<
+      int Function(int, double, double, int, int, double,
+          ffi.Pointer<ffi.UnsignedInt>)>();
 
   /// Stop already loaded sound identified by [handle] and clear it.
   ///
@@ -436,6 +449,19 @@ class FlutterSoLoudFfi {
   late final _disposeAllSound =
       _disposeAllSoundPtr.asFunction<void Function()>();
 
+  /// Query whether a sound is set to loop.
+  ///
+  /// [handle]
+  /// Returns true if flagged for looping.
+  bool getLooping(int handle) {
+    return _getLooping(handle) == 1;
+  }
+
+  late final _getLoopingPtr =
+      _lookup<ffi.NativeFunction<ffi.Int Function(ffi.UnsignedInt)>>(
+          'getLooping');
+  late final _getLooping = _getLoopingPtr.asFunction<int Function(int)>();
+
   /// This function can be used to set a sample to play on repeat,
   /// instead of just playing once
   ///
@@ -450,6 +476,36 @@ class FlutterSoLoudFfi {
     'setLooping',
   );
   late final _setLooping = _setLoopingPtr.asFunction<void Function(int, int)>();
+
+  /// Get sound loop point value.
+  ///
+  /// [handle]
+  /// Returns the time in seconds.
+  double getLoopPoint(int handle) {
+    return _getLoopPoint(handle);
+  }
+
+  late final _getLoopPointPtr =
+      _lookup<ffi.NativeFunction<ffi.Double Function(ffi.UnsignedInt)>>(
+          'getLoopPoint');
+  late final _getLoopPoint =
+      _getLoopPointPtr.asFunction<double Function(int)>();
+
+  /// Set sound loop point value.
+  ///
+  /// [handle]
+  /// [time] in seconds.
+  void setLoopPoint(SoundHandle handle, double time) {
+    _setLoopPoint(handle.id, time);
+  }
+
+  late final _setLoopPointPtr = _lookup<
+          ffi.NativeFunction<ffi.Void Function(ffi.UnsignedInt, ffi.Double)>>(
+      'setLoopPoint');
+  late final _setLoopPoint =
+      _setLoopPointPtr.asFunction<void Function(int, double)>();
+
+  /// TODO(@alnitak): implement Soloud.getLoopCount() also?
 
   /// Enable or disable visualization
   ///
@@ -954,10 +1010,17 @@ class FlutterSoLoudFfi {
   /// 3D audio methods
   /////////////////////////////////////////
 
-  /// play3d() is the 3d version of the play() call.
+  /// play3d() is the 3d version of the play() call
   ///
-  /// Returns the handle of the sound, [SoundHandle.error] if error
-  SoundHandle play3d(
+  /// [posX], [posY], [posZ] are the audio source position coordinates.
+  /// [velX], [velY], [velZ] are the audio source velocity.
+  /// [looping] whether to start the sound in looping state.
+  /// [loopingStartAt] If looping is enabled, the loop point is, by default,
+  /// the start of the stream. The loop start point can be set with this parameter, and
+  /// current loop point can be queried with [getLoopingPoint] and
+  /// changed by [setLoopingPoint].
+  /// Returns the handle of the sound, 0 if error
+  ({PlayerErrors error, SoundHandle newHandle}) play3d(
     SoundHash soundHash,
     double posX,
     double posY,
@@ -967,8 +1030,12 @@ class FlutterSoLoudFfi {
     double velZ = 0,
     double volume = 1,
     bool paused = false,
+    bool looping = false,
+    double loopingStartAt = 0,
   }) {
-    final handleId = _play3d(
+    // ignore: omit_local_variable_types
+    final ffi.Pointer<ffi.UnsignedInt> handle = calloc();
+    final e = _play3d(
       soundHash.hash,
       posX,
       posY,
@@ -978,38 +1045,34 @@ class FlutterSoLoudFfi {
       velZ,
       volume,
       paused ? 1 : 0,
+      looping ? 1 : 0,
+      loopingStartAt,
+      handle,
     );
-    if (handleId == 0) {
-      return SoundHandle.error();
-    }
-    return SoundHandle(handleId);
+    final ret =
+        (error: PlayerErrors.values[e], newHandle: SoundHandle(handle.value));
+    calloc.free(handle);
+    return ret;
   }
 
   late final _play3dPtr = _lookup<
       ffi.NativeFunction<
           ffi.UnsignedInt Function(
-            ffi.UnsignedInt,
-            ffi.Float,
-            ffi.Float,
-            ffi.Float,
-            ffi.Float,
-            ffi.Float,
-            ffi.Float,
-            ffi.Float,
-            ffi.Int,
-          )>>('play3d');
+              ffi.UnsignedInt,
+              ffi.Float,
+              ffi.Float,
+              ffi.Float,
+              ffi.Float,
+              ffi.Float,
+              ffi.Float,
+              ffi.Float,
+              ffi.Int,
+              ffi.Int,
+              ffi.Double,
+              ffi.Pointer<ffi.UnsignedInt>)>>('play3d');
   late final _play3d = _play3dPtr.asFunction<
-      int Function(
-        int,
-        double,
-        double,
-        double,
-        double,
-        double,
-        double,
-        double,
-        int,
-      )>();
+      int Function(int, double, double, double, double, double, double, double,
+          int, int, double, ffi.Pointer<ffi.UnsignedInt>)>();
 
   /// Since SoLoud has no knowledge of the scale of your coordinates,
   /// you may need to adjust the speed of sound for these effects