+++ lib/src/enums.dart
@@ -46,7 +46,8 @@ enum CaptureErrors {
   String toString() => 'CaptureErrors.$name ($_asSentence)';
 }
 
-/// Possible player errors
+/// Possible player errors.
+/// New values must be enumerated at the bottom
 enum PlayerErrors {
   /// No error
   noError,
@@ -85,6 +86,12 @@ enum PlayerErrors {
   /// Player not initialized
   backendNotInited,
 
+  /// Filter not found
+  filterNotFound,
+
+  /// asking for wave and FFT is not enabled
+  visualizationNotEnabled,
+
   /// Audio isolate already started
   @Deprecated('Use multipleInitialization instead')
   isolateAlreadyStarted,
@@ -110,7 +117,10 @@ enum PlayerErrors {
   filterNotFound,
 
   /// asking for wave and FFT is not enabled
-  visualizationNotEnabled;
+  visualizationNotEnabled,
+
+  /// Asset was found but for some reason couldn't be loaded.
+  assetLoadFailed;
 
   /// Returns a human-friendly sentence describing the error.
   String get _asSentence {
@@ -164,6 +174,10 @@ enum PlayerErrors {
       case PlayerErrors.visualizationNotEnabled:
         return 'Asking for audio data is not enabled! Please use '
             '`setVisualizationEnabled(true);` to enable!';
+      case PlayerErrors.assetLoadFailed:
+        return "Asset was found but for some reason couldn't be loaded. "
+            'This could be a problem with the temporary directory into which '
+            'the asset is being copied.';
     }
   }