+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#### 2.0.0-pre.X
+- added `looping` and `loopingStartAt` properties to `SoLoud.play()` and `SoLoud.play3d()`.
+- added `SoLoud.getLooping()` to retrieve the looping state of a sound.
+- added `SoLoud.getLoopPoint()` and `SoLoud.setLoopPoint()` to get and set the looping start position of a sound.
+
 #### 2.0.0-pre.0 (11 Mar 2024)
 - added `bool SoLoud.getVisualizationEnabled()` to get the current state of the visualization.
 - added `mode` property to `SoLoud.loadFile()` and `SoloudTools.loadFrom*` to prevent to load the whole audio data into memory:
@@ -52,6 +57,12 @@
 - All fields of `SoundProps` marked `final`. This is a breaking change
   but unlikely to have effect (as most users hopefully don't assign
   to these fields).
+- New methods `SoLoud.loadAsset()` and `SoLoud.loadUrl()` to load audio
+  from assets and URLs, respectively. These replace the old 
+  `SoloudTools.loadFrom*` methods (which are now deprecated).
+    - The new methods also correctly invalidate the temporary files
+      (for example, when an asset changes between versions of the app,
+      we don't want to play the old file).
 
 #### 1.2.5 (2 Mar 2024)
 - updated mp3, flac and wav decoders